Extract socketless client and add specs
